+++ frontend/src/pages/dashboard/components/calendar/index.ts
@@ -1,1 +1,5 @@
 export * from './calendar'
+export * from './components'
+export * from './date-picker'
+export * from './reducer'
+export * from './utils'
